--- plugin/arcconf/__init__.py
@@ -1,1 +1,0 @@
-from lsm.plugin.arcconf.arcconf import Arcconf
