+++ plugin/arcconf_plugin/__init__.py
@@ -1,0 +1,1 @@
+from arcconf_plugin.arcconf import Arcconf
